Refactor TableContainer: Rename and Relocate Component to TableGrid
