+++ src/components/examples/header-groups.tsx
@@ -5,7 +5,7 @@ import dummyData from "@/data/dummy.json";
 //   TableGrid,
 //   Column,
 // } from "@shakibdshy/react-tablegrid";
-import { TableContainer as TableGrid } from "@/components/containers/table-container/table-container";
+import { TableGrid } from "@/components/containers/table-grid/table-grid";
 import { createColumnHelper } from "@/utils/column-helper";
 import type { Column } from "@/types/column.types";
